# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evolution API Proxy",
    "version": "1.0.0",
    "description": "Proxy backend for Evolution API enabling ChatGPT Actions integration for WhatsApp automation",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Evolution API key for authentication"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SendTextMessageRequest": {
        "type": "object",
        "required": ["number", "text"],
        "properties": {
          "number": {
            "type": "string",
            "description": "Recipient phone number (international format without +)",
            "example": "5511999999999"
          },
          "text": {
            "type": "string",
            "description": "Message text content",
            "example": "Hello from Evolution API!"
          },
          "delay": {
            "type": "integer",
            "description": "Delay in milliseconds before sending",
            "example": 1000
          }
        }
      },
      "SendMediaMessageRequest": {
        "type": "object",
        "required": ["number", "mediatype", "media"],
        "properties": {
          "number": {
            "type": "string",
            "description": "Recipient phone number",
            "example": "5511999999999"
          },
          "mediatype": {
            "type": "string",
            "enum": ["image", "video", "audio", "document"],
            "description": "Type of media to send",
            "example": "image"
          },
          "media": {
            "type": "string",
            "format": "uri",
            "description": "URL of the media file",
            "example": "https://example.com/image.jpg"
          },
          "caption": {
            "type": "string",
            "description": "Caption for the media",
            "example": "Check out this image!"
          },
          "fileName": {
            "type": "string",
            "description": "File name for the media",
            "example": "photo.jpg"
          }
        }
      },
      "SendListMessageRequest": {
        "type": "object",
        "required": ["number", "title", "buttonText", "sections"],
        "properties": {
          "number": {
            "type": "string",
            "example": "5511999999999"
          },
          "title": {
            "type": "string",
            "example": "Choose an option"
          },
          "description": {
            "type": "string",
            "example": "Please select one of the following options"
          },
          "buttonText": {
            "type": "string",
            "example": "View Options"
          },
          "footerText": {
            "type": "string",
            "example": "Powered by Evolution API"
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Section 1"
                },
                "rows": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "Option 1"
                      },
                      "description": {
                        "type": "string",
                        "example": "Description for option 1"
                      },
                      "rowId": {
                        "type": "string",
                        "example": "option_1"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "SendButtonMessageRequest": {
        "type": "object",
        "required": ["number", "title", "buttons"],
        "properties": {
          "number": {
            "type": "string",
            "example": "5511999999999"
          },
          "title": {
            "type": "string",
            "example": "Choose an action"
          },
          "description": {
            "type": "string",
            "example": "Please select one of the buttons below"
          },
          "footer": {
            "type": "string",
            "example": "Powered by Evolution API"
          },
          "buttons": {
            "type": "array",
            "maxItems": 3,
            "items": {
              "type": "object",
              "properties": {
                "buttonId": {
                  "type": "string",
                  "example": "btn_1"
                },
                "buttonText": {
                  "type": "object",
                  "properties": {
                    "displayText": {
                      "type": "string",
                      "example": "Click Here"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "SetWebhookRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Webhook URL to receive events",
            "example": "https://your-webhook-url.com/webhook"
          },
          "webhook_by_events": {
            "type": "boolean",
            "description": "Send webhooks by event type",
            "example": true
          },
          "webhook_base64": {
            "type": "boolean",
            "description": "Include base64 media in webhooks",
            "example": false
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of events to listen to",
            "example": ["messages.upsert", "messages.update"]
          }
        }
      }
    }
  },
  "paths": {
...
    "/instances/{instanceName}/messages/send": {
      "post": {
        "summary": "Send text message",
        "description": "Send a text message to a WhatsApp number",
        "operationId": "sendTextMessage",
        "tags": ["Message Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/messages/sendMedia": {
      "post": {
        "summary": "Send media message",
        "description": "Send a media message (image, video, audio, or document)",
        "operationId": "sendMediaMessage",
        "tags": ["Message Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Media message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/messages/list": {
      "get": {
        "summary": "List messages from chat",
        "description": "Get messages from a specific chat",
        "operationId": "listMessages",
        "tags": ["Message Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chat ID to retrieve messages from"
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/messages/sendList": {
      "post": {
        "summary": "Send list message",
        "description": "Send an interactive list message",
        "operationId": "sendListMessage",
        "tags": ["Message Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendListMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instanceName}/messages/sendButtons": {
      "post": {
        "summary": "Send button message",
        "description": "Send an interactive button message (max 3 buttons)",
        "operationId": "sendButtonMessage",
        "tags": ["Message Operations"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendButtonMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Button message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/instances/{instanceName}/webhook": {
      "get": {
        "summary": "Get webhook configuration",
        "description": "Get current webhook configuration for instance",
        "operationId": "getWebhook",
        "tags": ["Webhook Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set webhook URL",
        "description": "Configure webhook URL for receiving events",
        "operationId": "setWebhook",
        "tags": ["Webhook Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetWebhookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook configured successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove webhook",
        "description": "Remove webhook configuration",
        "operationId": "deleteWebhook",
        "tags": ["Webhook Management"],
        "parameters": [
          {
            "name": "instanceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}